release: v1.5.3

+++ src/core/empty-directory.js
@@ -1,22 +1,37 @@
 const fs = require('fs');
 const path = require('path');
 
-function emptyDirectory(directoryPath)
+function emptyDirectory(directoryPath, targetDirectory = null, exceptionList = [])
 {
     const files = fs.readdirSync(directoryPath);
+
+    if (targetDirectory)
+    {
+        const targetFiles = fs.readdirSync(targetDirectory);
+
+        for (const file of targetFiles)
+        {
+            const filePath = path.join(directoryPath, file);
+            exceptionList.push(filePath);
+        }
+    }
 
     for (const file of files)
     {
         const filePath = path.join(directoryPath, file);
         const stat = fs.statSync(filePath);
 
-        if (stat.isDirectory())
+        if (!exceptionList.includes(filePath))
         {
-            emptyDirectory(filePath);
-            fs.rmdirSync(filePath);
-        } else
-        {
-            fs.unlinkSync(filePath);
+            if (stat.isDirectory())
+            {
+                emptyDirectory(filePath, null, exceptionList);
+                fs.rmdirSync(filePath);
+            }
+            else
+            {
+                fs.unlinkSync(filePath);
+            }
         }
     }
 }
